+++ src/bk/__init__.py
@@ -1,3 +1,3 @@
-"""GPUbk booking package."""
+"""GPUBK booking package."""
 
 __version__ = "0.2.0rc1"